Labview Code
The interface between the Keithley 6221 current source, the Keithey 2182A nanovoltmeter, and Labview is a National Instruments GIPB-USB. I only had access to one, so the 1GIPB vi prompts the user to manually set the current, while the Switch GIPB vi allows the user to set the current once via Labview and then breaks for the user to manually switch the GIPB-USB to the nanovoltmeter.
The code here sets up these two instruments, then proceeds to send current across two leads, measuring the voltage across the remaining two and recording the value. It will do this for the 8 resistivity configurations with 100 trials. The user will then be prompted to add a positive magnetic field, and the vi will take 100 more measurements of each of the 4 configurations. The user will switch to a negative magnetic field, and 4 more measurements will be taken. Labview writes all of this data to a .txt/.csv file (example of format will also be included as reference).