Labview Master Vi
Used with the configuration subvi's to control the current source and nanovoltmeter and write the measurements to a .csv/.txt file.
The B fields first vi serves the same purpose and writes files the same way. The only difference is that the manual switching of the magnetic fields happens first.